migrate some methods into ApiTool that are only used by it, respond to some feedback
